requirements.txt
discord.py==2.3.1
motor==3.3.1
pymongo==4.5.0
python-dotenv==1.0.0
asyncio==3.4.3
psutil==5.9.5
aiohttp==3.8.5
python-dateutil==2.8.2